dawn_native: Add validation for binding number
The WGSL spec states that the binding number is a u16 literal. This
change adds validation for this.

See: https://gpuweb.github.io/gpuweb/wgsl/#attribute-binding
See: https://github.com/gpuweb/gpuweb/issues/1783

+++ src/dawn_native/IntegerTypes.h
@@ -22,7 +22,7 @@
 
 namespace dawn_native {
     // Binding numbers in the shader and BindGroup/BindGroupLayoutDescriptors
-    using BindingNumber = TypedInteger<struct BindingNumberT, uint32_t>;
+    using BindingNumber = TypedInteger<struct BindingNumberT, uint16_t>;
 
     // Binding numbers get mapped to a packed range of indices
     using BindingIndex = TypedInteger<struct BindingIndexT, uint32_t>;
